+++ src/missions_lib/greet_all.py
@@ -3,36 +3,46 @@
 # that it sees who are known
 
 import rospy
-from std_msgs.msg import Empty
 from smach import State
+from actionlib_msgs.msg import GoalStatus
+from std_msgs.msg import Empty
+from servo_msgs.msg import pan_tilt
 
-# MOVE_CAMERA_GRT State
-class MoveCameraGeeting(State):
+
+# PREPARE_FOR_MOVEMENT_GRT State
+class PrepareMovementGeeting(State):
     def __init__(self, helper_obj):
-        State.__init__(self, outcomes=['complete','awaiting_movement'],
+        State.__init__(self, outcomes=['complete','move'],
                        input_keys=['start_in'],
-                       output_keys=['start_out','seen_dict'])
+                       output_keys=['start_out','seen_dict','user_data_absolute','user_data_pan','user_data_tilt'])
         self.__helper_obj = helper_obj
                                                              
     def execute(self, userdata):
+        position_request = pan_tilt()
+
         # Is this the start of a new mission
         if userdata.start_in == True:
             userdata.start_out = False
             # clear the seen dictionary
             userdata.seen_dict = {}
             scan_complete = False
-            # Set the camera position to start position (pan min and tilt min)
-            self.__helper_obj.CameraToStartPos()
+            # get the camera start position (pan min and tilt min)
+            position_request = self.__helper_obj.CameraToStartPos()
         else:            
-            scan_complete = self.__helper_obj.CameraToNextPos()
+            scan_complete, position_request = self.__helper_obj.CameraToNextPos()
         
+        # Set up user data that will be used for goal in next state if not complete
+        userdata.user_data_absolute = True
+        userdata.user_data_pan = position_request.pan
+        userdata.user_data_tilt = position_request.tilt
+
         if scan_complete == True:
             next_outcome = 'complete'
         else:
-            next_outcome = 'awaiting_movement'
+            next_outcome = 'move'
         
-        return next_outcome                                                            
-
+        return next_outcome 
+                                                           
 # Greeting State
 class Greeting(State):
     def __init__(self, helper_obj):
@@ -74,15 +84,12 @@ class GreetingHelper():
         self.__scan_request_pub = rospy.Publisher('face_recognition_node/start',
                                                   Empty, queue_size=1)
 
-    def MovementCompleteGrt(self, userdata, msg):
-        # Camera movement is complete
-
-        # Request the face recognition
-        face_recognition_request = Empty()
-        self.__scan_request_pub.publish(face_recognition_request)         
-        
-        # Returning False means the state transition will follow the invalid outcome
-        return False
+    def MovementCompleteCB(self, userdata, status, result):
+        if status == GoalStatus.SUCCEEDED:
+            # Camera movement is complete
+            # Request the face recognition
+            face_recognition_request = Empty()
+            self.__scan_request_pub.publish(face_recognition_request)    
         
     def AnalysisCompleteGrt(self, userdata, msg):
         # Analysis for face recognition is complete